sm8150-common: DeviceSettings: Add dolby atmos disable switch
we can get rid of it by disabling the service component and stopping the
service.
Enable works on the fly by just stopping starting audio.
Disable seems to require a reboot even if we stop service and
everything.

Adapted slightly to fit in our basic DeviceSettings.

Originally authored and commited by: Max Weninger
<max.weninger@gmail.com>

Change-Id: I1ad8b427848ee70b39e07101a89fc121d3f2d82f
Signed-off-by: AnierinB <anierinb@evolution-x.org>